+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [6.0.0]
 
 ### Added
 
 - autoload-dev to composer file for OXID devs to test locally
 - phpstan
+- GitHub actions for building project
+- Validation on first and last name according to EasyCredit model
 
 ### Changed
 
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Change log from README.md
+- Static call in [EasyCreditPayment](Core/Domain/EasyCreditPayment.php)
 
 ### Fixed